custom sentences are also optional
