+++ Optional View Assist custom sentences/Broadcast/blueprint-broadcast.yaml
