+++ Optional View Assist custom sentences/Broadcast/broadcast.yaml
