+++ Optional View Assist custom sentences/Device Functions/README.md
