+++ Optional View Assist custom sentences/Device Functions/blueprint-devicefunctions.yaml
