+++ Optional View Assist custom sentences/Get Sports Scores/README.md
