+++ Optional View Assist custom sentences/Get Sports Scores/blueprint-getsportsscores.yaml
