+++ Optional View Assist custom sentences/Hows the weather/README.md
