+++ Optional View Assist custom sentences/Hows the weather/blueprint-howstheweather.yaml
