+++ Optional View Assist custom sentences/List Management/README.md
