+++ Optional View Assist custom sentences/List Management/blueprint-listmanagement.yaml
