+++ Optional View Assist custom sentences/Search Wikipedia/blueprint-searchwikipedia.yaml
