+++ Optional View Assist custom sentences/Spell a Word/README.md
