+++ Optional View Assist custom sentences/Spell a Word/spellaword.yaml
