+++ Optional View Assist custom sentences/Tell Me a Joke/README.md
