+++ Optional View Assist custom sentences/Tell Me a Joke/blueprint-tellmeajoke.yaml
