+++ Optional View Assist custom sentences/Tell Me a Joke/viewassist-get_dadjoke.py
