+++ Optional View Assist custom sentences/Thank You/README.md
