+++ Optional View Assist custom sentences/Thank You/blueprint-thankyou.yaml
