+++ Optional View Assist custom sentences/Thank You/thankyou.yaml
